pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlx-week-4"
version = "0.1.0"
description = "MLX Week 4 Project"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx",
    "numpy",
    "matplotlib",
    "jupyter",
    "transformers>=4.53.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "nltk>=3.9.1",
    "rouge-score>=0.1.2",
    "huggingface-hub>=0.33.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8", 
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["notebooks"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
